Break out a GameInfo component
