+++ src/components/Game.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { Board } from "./Board";
+import { GameInfo } from "./GameInfo";
 
 function winner(squares) {
     const lines = [
@@ -31,7 +32,7 @@ export function Game() {
 
     const nextPlayer = () => moveNumber % 2 === 0 ? "X" : "O";
 
-    function handleClick(i) {
+    function handleSquareClick(i) {
         const subhistory = history.slice(0, moveNumber + 1);
         const current = subhistory[subhistory.length - 1];
         const squares = current.squares.slice();
@@ -40,9 +41,8 @@ export function Game() {
             return;
         }
 
-        squares[i] = nextPlayer();0
+        squares[i] = nextPlayer();
         setHistory(subhistory.concat([{squares: squares}]));
-
         setMoveNumber(moveNumber + 1);
     }
 
@@ -53,25 +53,10 @@ export function Game() {
         ? `Winner: ${w}`
         : `Next player: ${nextPlayer()}`;
 
-    const description = i => i > 0
-        ? `Go to move #${i}`
-        : "Go to start";
-
-    const moves = history.map((x, i) => {
-        return (
-            <li key={i.toString()}>
-                <button onClick={() => setMoveNumber(i)}>{description(i)}</button>
-            </li>
-        );
-    });
-
     return (
         <div className="game">
-            <Board squares={current.squares} onClick={handleClick} />
-            <div className="game-info">
-                <div>{status}</div>
-                <ol>{moves}</ol>
-            </div>
+            <Board squares={current.squares} onClick={handleSquareClick} />
+            <GameInfo status={status} history={history} setMoveNumber={setMoveNumber} />
         </div>
     );
 }
